+++ src/test/java/mb/minecraft/model/VillageTest.java
@@ -15,23 +15,23 @@ public class VillageTest {
 	@Test
 	public void testEntityBuilder() {
 		Village village = Village.builder()
-				.id( 10L )
+				.id( 10 )
 				.name("Montara")
 				.build();
 		assertNotNull( village );
 		assertNotNull( village.toString() );
-		assertEquals( 10L, village.getId().longValue() );
+		assertEquals( 10, village.getId().intValue() );
 		assertEquals( "Montara", village.getName() );
 	}
 
 	@Test
 	public void testEntitySetters() {
 		Village village = new Village();
-		village.setId( 10L );
+		village.setId( 10 );
 		village.setName("Montara");
 		assertNotNull( village );
 		assertNotNull( village.toString() );
-		assertEquals( 10L, village.getId().longValue() );
+		assertEquals( 10, village.getId().intValue() );
 		assertEquals( "Montara", village.getName() );
 	}
